get rid of waterspout and rename query_measurements2
